--- src/jjjexperiment/v_min_input/__init__.py
@@ -1,2 +1,0 @@
-from .section4_2_a import *
-from .inputs import *
